voicevox/requirements.txt
dify_plugin
ffmpeg
httpx
pydub
pyyaml
